requirements.txt
appdirs==1.4.4
asgiref==3.7.2
distlib==0.3.7
Django==4.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
filelock==3.12.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.4.1